# swagger.yaml
openapi: 3.0.0
info:
  title: Music Playlist Manager API
  description: A comprehensive music playlist management system with JWT authentication, file uploads, and Redis caching
  version: 1.0.0
  contact:
    name: Ajang Deng
    url: https://github.com/Ajang-Deng98/fullstack_assignment_group5
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Song:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        duration:
          type: number
        genre:
          type: string
        filePath:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /songs:
    post:
      tags:
        - Songs
      summary: Create a new song
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                artist:
                  type: string
                album:
                  type: string
                duration:
                  type: number
                genre:
                  type: string
                audioFile:
                  type: string
                  format: binary
      responses:
        201:
          description: Song created successfully
        401:
          description: Unauthorized

    get:
      tags:
        - Songs
      summary: Get all songs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Songs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Song'
                  pagination:
                    type: object
...
  /playlists:
    post:
      tags:
        - Playlists
      summary: Create a new playlist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                userId:
                  type: string
                isPublic:
                  type: boolean
      responses:
        201:
          description: Playlist created successfully
        401:
          description: Unauthorized

    get:
      tags:
        - Playlists
      summary: Get all playlists
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Playlists retrieved successfully
...